requirements.txt
astroid==3.2.4
autopep8==2.3.1
certifi==2024.8.30
chardet==5.2.0
defusedxml==0.6.0
Django==5.1.1
django-allauth==64.2.1
django-cors-headers==4.4.0
dj-rest-auth==6.0.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.8
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.7.0
oauthlib==3.2.2
pycodestyle==2.12.1
pylint==3.2.7
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
typed-ast==1.5.5
urllib3==2.2.2
wrapt==1.16.0
whitenoise==6.7.0
djangorestframework-simplejwt==5.3.1